Rebased down the commits.
CHANGE LOG:

Switched back to the transform; grabbed the outputs from the yaml file.

this is it chief.

AWSTemplateFormationVersion -> AWSTemplateFormatVersion.

Cloudformation now says this is a valid template.

Diagram of simple site with route53 record set.

updated the s3 bucket for the site to have a name rather than the super long one since I forgot to list anything there.

Syntax error on the sub usage. Corrected.